### 2023-03-08 #5235 @editorialbot
[PRE REVIEW]: VaMPy: An Automated and Objective Pipeline for Modeling Vascular Geometries
**Submitting author:** @hkjeldsberg (<a href="http://orcid.org/0000-0002-7764-4248">Henrik Aasen Kjeldsberg</a>)
**Repository:** https://github.com/KVSlab/VaMPy
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.1
