+++ paper/paper.bib
@@ -1,6 +1,7 @@
 %==Ref1
 
 @article{valen2014mind,
+  doi = {10.3174/ajnr.a3793},
   title={Mind the gap: impact of computational fluid dynamics solution strategy on prediction of intracranial aneurysm hemodynamics and rupture status indicators},
   author={Valen-Sendstad, Kristian and Steinman, David A},
   journal={American Journal of Neuroradiology},
@@ -12,6 +13,7 @@
 }
 
 @article{khan2015narrowing,
+  doi = {10.3174/ajnr.a4263},
   title={Narrowing the expertise gap for predicting intracranial aneurysm hemodynamics: impact of solver numerics versus mesh and time-step resolution},
   author={Khan, Malek O and Valen-Sendstad, Kristian and Steinman, David A},
   journal={American Journal of Neuroradiology},
@@ -23,6 +25,7 @@
 }
 
 @article{valen2014high,
+  doi = {10.1016/j.jbiomech.2014.04.018},
   title={High-resolution computational fluid dynamics detects flow instabilities in the carotid siphon: implications for aneurysm initiation and rupture?},
   author={Valen-Sendstad, Kristian and Piccinelli, Marina and Steinman, David A},
   journal={Journal of biomechanics},
@@ -33,350 +36,8 @@
   publisher={Elsevier}
 }
 
-@article{bozzetto2016transitional,
-  title={Transitional flow in the venous side of patient-specific arteriovenous fistulae for hemodialysis},
-  author={Bozzetto, Michela and Ene-Iordache, Bogdan and Remuzzi, Andrea},
-  journal={Annals of biomedical engineering},
-  volume={44},
-  number={8},
-  pages={2388--2401},
-  year={2016},
-  publisher={Springer}
-}
-
-@article{bozzetto2022,
-  title={Turbulent-like arteriovenous fistula flows cause wall vibrations: a specific stimulus for stenosis formation?},
-  author={Bozzetto, Michela and Ring, Johannes and Remuzzi, Andrea and Valen-Sendstad, Kristian},
-  journal={Submitted to Annals of Biomedical Engineering},
-  year={2022},
-  publisher={Springer}
-}
-
-@article{Hirt1974,
-      author = "Hirt, CW and Amsden, AA and Cook, JL",
-       title = "An arbitrary {L}agrangian-{E}ulerian computing method for all flow speeds",
-        year = "1974",
-     journal = "J {C}omput {P}hys",
-      volume = "14",
-      number = "3",
-       pages = "227--253"
-}
-
-@article{bozzetto2022arteriovenous,
-  title={Arteriovenous fistula creation with VasQTM device: A feasibility study to reveal hemodynamic implications},
-  author={Bozzetto, Michela and Soliveri, Luca and Poloni, Sofia and Brambilla, Paolo and Curt{\`o}, Diego and Condemi, Giuseppina Carmela and Cefal{\`\i}, Pietro and Spina, Irene and Villa, Alessandro and Caroli, Anna and others},
-  journal={The Journal of Vascular Access},
-  pages={11297298221087160},
-  year={2022},
-  publisher={SAGE Publications Sage UK: London, England}
-}
-
-%==Ref2
-@article{Benson1992,
-      author = "Benson, DJ",
-       title = "Computational methods in {L}agrangian and {E}ulerian hydrocodes",
-        year = "1992",
-     journal = "Comput {M}ethod {A}ppl {M}",
-      volume = "99",
-      number = "2--3",
-       pages = "235--394"
-}
-
-%==Ref3
-@article{Dukowicz1984,
-      author = "Dukowicz, J",
-       title = "Conservative rezoning (remapping) for general quadrilateral meshes",
-        year = "1984",
[…]
