### 2023-03-09 #5235 @hkjeldsberg
Hi @Kevin-Mattheus-Moerman,

We would like to suggest reaching out to @brainstorm and @rlizzo to review our paper since they have reviewed a similar paper in JOSS, which VaMPY builds upon.

Additionally, we believe that @JaroslavHron, @ctdegroot, @capitalaslash, and [herbert.owen@bsc.es](mailto:herbert.owen@bsc.es) would be ideal candidates for reviewing this software/paper, given their expertise in CFD and/or FEM.
