### 2023-03-09 #5235 @Kevin-Mattheus-Moerman
@brainstorm @rlizzo @ElenaFagg @JaroslavHron, @ctdegroot, @capitalaslash @wouterpotters @aslakbergersen @ktbolt @thewtex @vijaysm @stefsmeets @prashjha would you be interested in reviewing this work for JOSS? The paper title is "VaMPy: An Automated and Objective Pipeline for Modeling Vascular Geometries", and it concerns a Python library for vascular modelling. As you might know, reviewing for JOSS means you evaluate/review the software and review a short paper. Links are given below. Let me know if you are interested. Thanks!

Software: https://github.com/KVSlab/VaMPy
Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.05235/joss.05235/10.21105.joss.05235.pdf
Information about reviewing for JOSS: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
