### 2023-03-20 #5278 @editorialbot
[REVIEW]: VaMPy: An Automated and Objective Pipeline for Modeling Vascular Geometries
**Submitting author:** @hkjeldsberg (<a href="http://orcid.org/0000-0002-7764-4248">Henrik Aasen Kjeldsberg</a>)
**Repository:** https://github.com/KVSlab/VaMPy
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.2
