### 2023-04-03 KVSlab/VaMPy#104 @stefsmeets
Installation instructions are unclear
Hi, I'm trying to install vampy on my machine, but I keep running into issues. Eventually I got it to work by trying this on Linux. So I'm wondering:

- Which platforms are supported?
- Which version of Python do I need?
- What libraries do I need to install to get this to work?

It is not clear in the readme which OS and python versions are supported. The documentation says it can be installed via pip, so I naively tried to install via pip (python 3.10) on Windows (`pip install vampy`), but this did not work. If there is a version available on pypi, it would be great if the project page could be linked in the readme.

Then I cloned the repo and installed via `pip install .`. This failed the tests, because I needed `dolfin`.

This time I followed the installation instructions, and it turns out that I was also not able to install this using `conda`.

```
PS > conda env update --file environment.yml --name your_environment
Solving environment: failed

ResolvePackageNotFound:
  - fenics
```

I then tried on linux, and this takes a very long time to install:

```
$ conda env update --file environment.yml --name your_environment
```

Eventually it did something and it seems to work.
