### 2023-04-03 #5278 @stefsmeets
I tried to install vampy as a naive user, and the installation instructions are not clear to me. See issue here:
https://github.com/KVSlab/VaMPy/issues/104
