### 2023-04-03 KVSlab/VaMPy#105 @stefsmeets
Issues running the examples in the documentation
Hi, I'm trying out vampy for the JOSS review (https://github.com/openjournals/joss-reviews/issues/5278). These are some issues I run into with the examples in the documentation. This is on Linux using Python 3.9 with the development version of vampy. I'm not sure if I did something wrong with the installation. Two error seem to be related to 'tetgen' inside of dolfin, and the other with some gzip related error.

Artery tutorial fails on tetgen part:

```console
$ vampy-mesh -m constant -el 0.3 -i c0001/model.vtp -bl False
...
Generating volume mesh
TetGen command line options: pq1.414000q10.000000q165.000000YsT1.000000e-08zQm
Internalerror in findedge():  Unable to find an edge in subface.
2023-03-29 06:23:15.179 ( 654.314s) [        FE135740]vtkvmtkTetGenWrapper.cx:433    ERR| vtkvmtkTetGenWrapper (0x55fb77b60c10): TetGen quit with an exception.
An error occurred during tetrahedralization. Will only output surface mesh.
Writing Dolfin file.
Segmentation fault
```

The artery example also fails with tetgen.

```console
$ vampy-mesh -i models/artery/artery.vtp -c 1.3
Generating volume mesh
TetGen command line options: pq1.414000q10.000000q165.000000YsT1.000000e-08zQm
Internalerror in findedge():  Unable to find an edge in subface.
2023-03-29 06:36:54.786 (  62.973s) [        F3E0B740]vtkvmtkTetGenWrapper.cx:433    ERR| vtkvmtkTetGenWrapper (0x55cadc27b6f0): TetGen quit with an exception.
An error occurred during tetrahedralization. Will only output surface mesh and boundary layer.
Assembling final mesh
2023-03-29 06:36:54.788 (  62.975s) [        F3E0B740]       vtkThreshold.cxx:84    WARN| vtkThreshold::ThresholdByUpper was deprecated for VTK 9.1 and will be removed in a future version.
Writing Dolfin file.
Segmentation fault
```

Finally, I get this error when running the simulation example:

```console
$ oasis NSfracStep problem=Artery mesh_path=../../../models/artery/artery.xml.gz save_solution_after_cycle=0
Importing problem module Artery:
/home/stef/python/vampy/src/vampy/simulation/Artery.py
=== Starting simulation for case: artery ===
{...}
Traceback (most recent call last):
  File "/home/stef/miniconda3/envs/your_environment/bin/oasis", line 8, in <module>
    sys.exit(main())
  File "/home/stef/miniconda3/envs/your_environment/lib/python3.10/site-packages/oasis/run_oasis.py", line 11, in main
    from oasis import NSfracStep
  File "/home/stef/miniconda3/envs/your_environment/lib/python3.10/site-packages/oasis/NSfracStep.py", line 60, in <module>
    vars().update(post_import_problem(**vars()))
  File "/home/stef/miniconda3/envs/your_environment/lib/python3.10/site-packages/oasis/problems/__init__.py", line 212, in post_import_problem
    mesh = mesh(**NS_parameters)
  File "/home/stef/python/vampy/src/vampy/simulation/Artery.py", line 82, in mesh
    mesh = Mesh(mesh_path)
RuntimeError: gzip error: iostream error
```
