### 2023-04-03 #5278 @stefsmeets
The paper is well written, and the introduction makes the goal of the project very clear to someone outside the field like myself. I like how the authors used a study on inter-lab variability as the motivation for the software. Overall a nice piece of work, and I fully subscribe to the idea of making complex research workflows available for a non-expert audience.

- I had some issues installing the software (see https://github.com/KVSlab/VaMPy/issues/104). I naively installed on Windows, but this may not have been the authors intention. I think a list of requirements and dependencies in the readme and/or installation instructions would go a long way. In the readme, the authors state that the package can be installed via pip. To me this means `pip install vampy`, but it turns out this is not the case. I can imagine if I was a non-expert user, I would have difficulties to install this.

- I also ran into some issues verifying the functionality of the code. I copy-pasted some of the commands from the tutorial, and some returned with errors. It seemed a bit hit-or-miss to me. I made an issue for this here: https://github.com/KVSlab/VaMPy/issues/105. This meant I could not fully evaluate the functionality of the software.

 - On authorship: Both H. Kjedsberg and A. Bergersen made significant code contributions. To me it is unclear what the contributions from V. Valen-Sendstad are. Before I check the box, could you clarify this? https://joss.readthedocs.io/en/latest/submitting.html#authorship

- The tests take very long. Is this normal?

- From a user point-of-view, running through the tutorial, some of the models take a long time to compute. It would be great to have some progress bar to keep me in the loop for how long it will take to generate. Another thought is to run the tutorial with somewhat lower precision.

### Text

Below are some suggestions for improving the text.

For the summary I recommend swapping the first two lines to point out that his concerns medical research. As someone with a materials science background wall-shear stress is something completely different :-)

*Medical researchers commonly use image-based computational fluid dynamics (CFD) to study vascular pathology, but modeling choices and simulation results vary widely. Hemodynamic forces, such as wall shear stress (WSS), play a role in vessel wall adaption and remodeling. However, directly measuring these forces is challenging.*

16: has shown to become -> has become
19: extensively used -> used extensively
27: withinin -> within
32: variaiblity -> variability
38: reproduciable science
Figure 1: Text overlay is somewhat difficult to read. I'd recommend adding the labels above the images.
49: inlet -> inlets?
