### 2023-04-04 60b41f1
Update documentation with conda and docker installation pages
