+++ README.md
@@ -1,4 +1,5 @@
 # VaMPy - Vascular Modeling Pypeline
+
 _________________
 [![GPL-3.0](https://img.shields.io/github/license/hkjeldsberg/vampy)](LICENSE)
 [![codecov](https://codecov.io/gh/KVSlab/VaMPy/branch/master/graph/badge.svg?token=M2NMX6HOSZ)](https://codecov.io/gh/KVSlab/VaMPy)
@@ -17,22 +18,31 @@ _________________
 
 Description
 -----------
-The Vascular Modeling Pypeline (VaMPy) is a collection of fully automated scripts used to prepare, run, and analyze cardiac and atrial morphologies.  This includes pre-processing scripts for meshing and probe sampling, two [Oasis](https://github.com/mikaem/Oasis) problem files for simulating flow in the [internal carotid artery](https://en.wikipedia.org/wiki/Internal_carotid_artery) and the [left atrium](https://en.wikipedia.org/wiki/Atrium_(heart)), and a variety of post-processing scripts for computing WSS-based metrics, more advanced turbulence metrics, and a variety of morphological parameters in patient-specific geometries. 
+The Vascular Modeling Pypeline (VaMPy) is a collection of fully automated scripts used to prepare, run, and analyze
+cardiac and atrial morphologies. This includes pre-processing scripts for meshing and probe sampling,
+two [Oasis](https://github.com/mikaem/Oasis) problem files for simulating flow in
+the [internal carotid artery](https://en.wikipedia.org/wiki/Internal_carotid_artery) and
+the [left atrium](https://en.wikipedia.org/wiki/Atrium_(heart)), and a variety of post-processing scripts for computing
+WSS-based metrics, more advanced turbulence metrics, and a variety of morphological parameters in patient-specific
+geometries.
 
 Installation
 ------------
-VaMPy and its dependencies can be installed with `conda` on Linux and macOS as explained [here](https://kvslab.github.io/VaMPy/conda.html).
-VaMPy can also be installed and run through its latest `Docker` image supported by Windows, Linux, and macOS, and explained [here](https://kvslab.github.io/VaMPy/docker.html).
+VaMPy is a Python package for Python >= 3.8, with main dependencies to [morphMan](https://github.com/KVSlab/morphMan)
+and [Oasis](https://github.com/mikaem/Oasis). VaMPy and its dependencies can be installed with `conda` on Linux and
+macOS as explained [here](https://kvslab.github.io/VaMPy/conda.html). The package can also be installed and run through
+its latest `Docker` image supported by Windows, Linux, and macOS, and
+explained [here](https://kvslab.github.io/VaMPy/docker.html).
 
 Documentation
 -------------
-VaMPy's documentation is hosted [here](https://kvslab.github.io/VaMPy).
-This includes two [tutorials](https://kvslab.github.io/VaMPy/tutorials.html), meant to guide the user through the basic steps of performing a computational fluid dynamic simulation in a vascular body.
+VaMPy's documentation is hosted [here](https://kvslab.github.io/VaMPy). This includes
+two [tutorials](https://kvslab.github.io/VaMPy/tutorials.html), meant to guide the user through the basic steps of
+performing a computational fluid dynamic simulation in a vascular body.
 
 Licence
 -------
-VaMPy is licensed under the GNU GPL, version 3 or (at your option) any
-later version.
+VaMPy is licensed under the GNU GPL, version 3 or (at your option) any later version.
 
 VaMPy is Copyright (2018-2023) by the authors.
 
