### 2023-04-04 KVSlab/VaMPy#104 @hkjeldsberg
Dear @stefsmeets ,

Thank you for highlighting this issue. It sounds like there were some trouble installing vampy on your machine. I have tried to address most of your comments through PR #107.
In summary, here are some key points regarding the installation of Vampy:

* Platform Support: Vampy can be installed on any operating system that supports Python, including Windows, macOS, and Linux.  On Linux and macOS, Vampy can be installed using conda or through the docker image. On Windows, it is a bit more complicated due to the FEniCS dependency. We recommend using the docker image for Windows, although there are workarounds for installing FEniCS.

* Python Version: Vampy requires Python 3.8 or higher.

* Required Libraries: The major dependencies for Vampy are FEniCS, Oasis, and morphMan. Additionally, there are some minor dependencies such as paramiko, cppimport, and matplotlib, which can be installed using either pip or conda.

I hope this, and the changes to the documentation in PR #107 will clarify the installation process of VaMPy.

Best regards,
Henrik
