+++ _sources/installation.md
@@ -1,84 +1,29 @@
 # Installation
 
-## Installing VaMPy using `conda`
+`VaMPy` is a pure Python package that combines the morphological processing tools of
+[morphMan](https://github.com/KVSlab/morphMan), the computational fluid dynamics
+solver [Oasis](https://github.com/mikaem/Oasis), and the finite element computing
+platform [FEniCS](https://fenicsproject.org/) into a powerful and user-friendly computational fluid dynamics pipeline.
 
-We recommend installing `VaMPy` and its dependencies through `conda`.  
-Start by cloning into the repository:
+## Dependencies
 
-``` console
-$ git clone https://github.com/KVSLab/VaMPy.git
-$ cd VaMPy
-```
+The dependencies of `VaMPy` are:
 
-Then, using the ``environment.yml`` file in the root of the repository, you can call:
+* Python >= 3.8
+* FEniCS >= 2018.1
+* morphMan >= 1.2
+* Oasis >= 2018.1
+* paramiko >= 3.0
+* cppimport >= 22.8.2
 
-``` console
-$ conda env update --file environment.yml --name your_environment
-```
+## Installing with `conda`
 
-Next, can now activate your environment by running::
+To install `VaMPy` and all its dependencies to a *local environment*, we recommend using `conda`.
+Instructions for installing `VaMPy`
+with `conda` can be found [here](install:conda).
 
-``` console
-$ conda activate your_environment
-```
+## Installing with Docker
 
-Finally, install `VaMPy` inside your `conda` environment using `pip`:
+To install `VaMPy` and all its dependencies to an *isolated environment*, we recommend using the dedicated Docker
+container. Instructions for installing `VaMPy` with Docker can be found [here](install:docker).
 
-``` console
-$ python3 -m pip install .
-```
-
-Now you are all set, and can start using the Vascular Modeling Pypeline.
-
-## Editable installations
-
-If you want to make changes to any of the underlying packages, you should remove them from the `environment.yml` file,
-and install them from source, as described for `Oasis` in the next section.
-
-### Installing Oasis
-
-To install an editable version of [Oasis](https://github.com/mikaem/Oasis) on your machine, run the following commands
-inside your conda environment:
-
-``` console
-$ git clone https://github.com/mikaem/Oasis
-$ pip install cppimport
-$ pip install --editable Oasis
-```
-
-The ``--editable`` flag installs the project in editable mode meaning that any changes to the original package will be
-reflected directly in your environment.
-
-## Running VaMPy using Docker
-
-A `Dockerfile` is supplied in the root directory of the repository, which can build a docker-image with all dependencies
-installed. The Docker-image can be built with the following command:
-
-``` console
-$ docker build -t name_of_image -f docker/Dockerfile .
-```
-
-A Docker-container can then be started with the following command:
-
-``` console
-$ docker run -ti --network=host -e DISPLAY=$DISPLAY -v /tmp/.X11-unix:/tmp/.X11-unix -v $(pwd):/root/shared -w /root/shared --rm --shm-size=512m name_of_image
-```
-
-To run the VaMPy GUI, you need to call:
-
-``` console
-$ xhost +local:root
-```
-
-on your system before running the scripts.
-
-### Note on Docker
-
-Remember to call:
-
-``` console
-xhost -local:root
-```
-
-on the host system when you are done running the Docker container.
-    
