### 2023-04-04 #5278 @hkjeldsberg
Hi @stefsmeets,

First of all, thank you for reviewing our software, that is greatly appreciated! Also, thank you for your detailed comment and suggestions. I'll try to address your points below:

* Installation Issues: Thank you for bringing this to our attention. Based on your input we realized that parts of the instructions were not clear enough. In response, we have updated the installation instructions in PR kvslab/vampy#107, and hope this answers some of the addressed issues.

* Verification Issues: We appreciate you bringing this to our attention as well. We believe the issue may be related to a well known issue regarding compression of the volumetric mesh, and have made a hotfix to bypass this issue in PR kvslab/vampy#108. Hopefully, this should improve your experience with Vampy.

* Authorship: We apologize for the confusion regarding the authorship. In short, it is Dr. Valen-Sendstad’s scripts that Vampy are partly based on, and he has actively contributed to the conceptualization and design of the software throughout its development. Therefore, we believe that Dr. Valen-Sendstad’s contributions are sufficient for co-authorship.

* Tests: Thank you for pointing this out. Based on your comment we noticed that the test data complexity was much higher than anticipated, so by reducing this we were able to reduce the total test time drastically. The fix was implemented in PR kvslab/vampy#109.

* Progress bar / tutorial precision: We appreciate your suggestion and will consider adding a progress bar where it can be beneficial. To keep the Artery tutorial high-resolution, we are limited to how coarse we can go, but for practical reasons there is is no real limit. Perhaps the wording should be changed from Tutorial to Demonstration, considerign it is more of a demonstration of the features of Vampy instead of a “user-friendly” demo.

*  Progress bar / tutorial precision: We appreciate your suggestion and will consider implementing a progress barit where it can be beneficial. Regarding the tutorial precision, we are limited to how coarse we can go while still maintaining the high-resolution nature of the Artery model. However, for practical purposes, there is no real limit to the level of precision we can use. In light of this, we will consider renaming the tutorial to a demonstration, as it is intended to showcase the features of Vampy rather than provide a fully user-friendly demo.

* Text suggestions: We appreciate and agree with your feedback on the text. Hence, in the latest version of the paper we have tried to incorporate your suggestions accordingly. Figure 1 has also been updated to reflect your suggestion.

Again, thank you for taking the time to provide us with feedback. And we hope that our responses have been helpful.
