### 2023-04-04 KVSlab/VaMPy#105 @hkjeldsberg
Dear @stefsmeets ,

Thank you for pointing out this issue. We believe the first two errors may be related to the issues we have experienced using the `vmtkMeshWriter` in the past due to errors in parsing in binary mode versus text mode. We tried to add a hotfix for this issue in #108, which has been merged to the master branch.

The last error experienced in `Artery.py` is likely caused by a defect in the volumetric mesh due to the segmentation fault above. I would recommend removing the `.vtu` and `.xml.gz` files, and see if re-meshing the geometry will help.

I hope this, and the changes made in PR #108 can help fix these issues.

Best regards,
Henrik
