### 2023-04-11 KVSlab/VaMPy#105 @stefsmeets
Thanks, the first 2 errors are resolved with #108

I still get the third error, but maybe this is more of an oasis issue than a vampy issue?
