### 2023-04-11 KVSlab/VaMPy#105 @stefsmeets
Actually, I still get the first 2 errors as well. I didn't realize the `.vtu`/`.xml.gz` files are cached. They are in the `.gitignore` file so I didn't realize they were not part of the repo.
