### 2023-04-13 KVSlab/VaMPy#105 @hkjeldsberg
Hi @stefsmeets,

Thanks for your patience. I investigated this further and think I have found out what is causing the issues you're experiencing.
Through the latest pull request (#110), I believe your two first errors should be fixed. I successfully tested this update thoroughly for both the `artery.vtp` model, and the `c0001` model on a Linux (Ubuntu), and tested this for multiple Python versions, including 3.8, 3.9, and 3.10.  Essentially the error is caused by some instability using the `vtkvmtkTetGenWrapper`, which I realise have been previously reported. Hopefully, with the changes in PR #110, you should now be able to mesh both models.

As for the third error, I think it is related to the first two errors. However, if the latest update does not seem to fix this issue I would try the following steps:

* Ensure that the `artery.xml.gz/artery.vtu` file generated by the `vampy-mesh` command is not corrupted or empty. One way to do this is to check that the `artery.vtu` file is *not* hollow using a visualization tool such as Paraview. It should be a dense mesh full of tetrahedral cells.
* Ensure that you have the latest versions of `oasis` and `vampy`.
* Try running the commands separately to see if that resolves the error. This could help isolate the issue and determine what is causing it.

I hope this, and the recent changes to `vampy` can help fix these issues.

Best regards,
Henrik
