### 2023-05-05 #5278 @JaroslavHron
This is a very nice collection of tools for the bio-CFD community. The article describes very well all the motivations and areas of possible use. It is well written.

- Among the dependencies, I would suggest to add or cite the `VMTK` library. This is probably a bit hidden inside the `morphMan` component. However, it seems to be substantially used in the mesh preparation part.
- I've had some problems installing the docker image, since the docker registry `https://github.com/KVSlab/VaMPy/pkgs/container/vampy` gives me a 404 error. I managed to install another image `docker pull hkjeldsberg/vampy` - is this the intended official image?
Unless this is some mistake on my part, it should be fixed in the docker installation instructions, as installing docker image might be the universal first choice for most users.
- Otherwise, I managed to do a source installation (since I have some of the prerequisites available) and then all the tests and instructions worked correctly.

@hkjeldsberg - I would recommend to fix the docker image installation instructions.
