### 2023-05-08 #5278 @hkjeldsberg
Dear @JaroslavHron,

Thank you for reviewing our VaMPy, we appreciate it!  We also appreciate your comments and suggestions, and have tried to address them:

* We agree with you regarding the point concerning the `VMTK` library. We'll make sure to mention it more clearly in our documentation, and have started by adding it to the dependencies in PR kvslab/vampy#111, which has been merged to the main branch.

* Apologize for the confusion with the Docker image installation. I looked into this, and it was simply the docker registry not being public. This should be fixed now, and you should have access through the same url (`https://github.com/KVSlab/VaMPy/pkgs/container/vampy`). The installation through Docker should now work as intended.

* The image located under `hkjeldsberg/vampy` is slightly deprecated, and this has now been deleted to avoid confusion.

Thanks again for pointing out these issues, and I hope that we were able to address them. If you have any other questions or suggestions, don't hesitate to reach out!
