### 2023-05-19 #5278 @Kevin-Mattheus-Moerman
@hkjeldsberg since the reviewers were happy to recommend acceptance we can now proceed to process this for acceptance in JOSS. Can you work on the above steps? :pointup:
Let me know what version number you'll use for your new tagged release and I'll update it here too.
