### 2023-05-19 #5278 @Kevin-Mattheus-Moerman
@hkjeldsberg I have just reviewed your paper, please work on the below as well:

- [ ] Please extend your affiliations to include the town/city and country (in full, no acronyms) as well.
- [ ]  Fix `Futhermore` -> `Furthermore`
- [ ] Fix `reproduciable` -> `reproducible`
- [ ] At the  moment the paper mixes American English (e.g. scrutinize, visualized, visualization, etc) with British English (e.g. visualised). I suggest you use one and revise accordingly.
- [ ] Several acronums, like CFD, OSI, RRT, TAWSS, and WSS are introduced several times. I suggest you check all acronyms, introduce them once and then use them as acronyms after.
