+++ src/vampy/automatedPreprocessing/moving_common.py
@@ -203,7 +203,7 @@ def project_displacement(clamp_boundaries, distance, folder_extended_surfaces, f
     """
     # Add extents to all surfaces
     extended_surfaces = sorted(
-        [f for f in listdir(folder_moved_surfaces) if f[:2] in ["LA", "Co", "Fu", "Ge", "Sm", "AF"]])
+        [f for f in listdir(folder_moved_surfaces) if f[:2] in ["LA", "Co", "Fu", "Ge", "Sm", "AF", "an"]])
     if not path.exists(folder_extended_surfaces):
         mkdir(folder_extended_surfaces)
 
